+++ app/controller/controller_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/leoleoasd/EduOJBackend/base"
 	"github.com/leoleoasd/EduOJBackend/base/config"
 	"github.com/leoleoasd/EduOJBackend/base/exit"
+	"github.com/leoleoasd/EduOJBackend/base/validator"
 	"github.com/leoleoasd/EduOJBackend/database"
 	"github.com/stretchr/testify/assert"
 	"io/ioutil"
@@ -115,6 +116,7 @@ server:
 	}
 
 	base.Echo = echo.New()
+	validator.InitValidator(base.Echo)
 	app.Register(base.Echo)
 	os.Exit(m.Run())
 }
